requirements.txt
requests==2.31.0
python-dotenv==1.0.1
psycopg2-binary==2.9.12
fastapi==0.111.1
uvicorn==0.23.2
